Added misc. features
Added output "data points" that tells how many times the run loop has executed.  Also added bounds on the values that can be picked for the user-chosen parameters such as max temperature and max current.  These max values can be changed later in the properties menu of each control on the front panel.