Fixed chart and added warning for max current.
Fixed lagend on chart so both inputs are labled.  Also, the the current sent to the power supply cannot be higher than the max current.  If the current calculated by the P controller is higher than the max current, a warning light tells the user that the curent is "pinned" against the maximum.