Read Temperature from Pyrometer
Temperature is now read from the analog output of the Mikron Pyrometer through the DAQ board. Units of T are now C instead of K.  If there is an error with the connection to the DAQ board the loop will stop running.  The manual input for detected temperature was removed.