Added an indicator for current temperature.
There is now an indicator on the front panel that displays the current temperature while the system is running.  This gets greyed out when the system is not running.  Also, the P, I , and D parameters now get written to teh text file along with all of the other parameters.  I and D both have placeholders until the actual PID algorithm is implemented.