Integral term added to control algorithm.
The control algorithm is now PI.  The proportional term is calculated by multiplying Kp with the error.  The integral term is calculated discreetly, by multiplying the average of the last two errors by dt, the amount of time it takes for one loop to complete.  These two terms are added together to give the current that is outputted by the power supply.  Also, the chart buffer length was increased to increase the amount of data in the chart and therefore slow down the rate at which the chart pans across.