Median filter replaced by running avg.
The point by point median filter that reduced the noise in the temperature data was replaced by a running average filter because of better observed performance.  The number of points was set to 50, but can be changed from the block diagram (maybe will add a user control later for this).  Also, the value for Ki now gets written into the text file.