Added Max Voltage indicator and removed power limit.
"Maximum Voltage" light was added, which tells the user that the voltage is at its max value and therefore the current cannot be raised any higher.  The power limit and max power indicator were both removed because of the redundancy of having voltage limits, current limits, and power limits.  In addition, having a power limit is misleading, because there is no mechanism to prevent the power supply from going over power.