Second loop created for reading temperature.
The temperature aquisition is now in a separate loop from the PID algorithm and the output to the power supply.  Each command to the power supply will take at least 10 ms, and because three commands need to be performed in each iteration of the loop, the loop will take at least 30 ms to iterate.  This does not really matter, because the temperature reading only changes once every 0.5 s, and I don't think the PID algorithm needs to run on the order of kHz anyway.  Because the temperature aquisition in a separate loop, it can repeat at a much faster rate, , without being limited by the time it takes to communicate with the power supply.  This way, there will be many more temperature data points over which a running average can be taken.  This will improve the signal to noise ratio of the temperature data.  The sample temperature is passed from one loop to the other using a local variable, as is the "stop" signal for both loops.