Arbitrary linear temperature schedule.
Now the user can use an arbitrary temperature schedule instead of just one temperature setpoint.  This was acomplished by using the setpoint VI included with the PID toolkit.  A subVI was created for the selection of the temperature setpoints so that it would be done in a separate window.  When the program first opens, the temperatures are reset.  Otherwise, they are just remembered from the last time the user selected them.  The UI of the temperature selection VI still needs to be finished, but it is functional.  It is important to note that the user is in charge of making sure that the time and temperature values are valid (e.g. time increases from each value to the next).