UIs edited.
The UIs for TPD_Controller and Temp_Controller were both edited.  The "change temperature" button in TPD_Controller now becomes disabled at the correct times.  In Temp_Controller, a title was added and some simple directions, including a warning that the user is responsible for checking that the temperature and time values are valid.  According to a quick test, if a user inputs values that are achronological, the run will stop immediately because the setpoint generator will not work.  This is better than the program just crashing.