Elapsed time now calculated by external VI.
The setpoint generator VI has an elapsed time output.  This is more accurate than the previous methos of calculating elapsed time, which was to use the "seconds since 1904" function.  The elapsed time from the setpoint generator VI is now wired through a shift register directly to the elapsed time array.