Measured Current displayed on chart.
The chart on the front panel now displays the measured Current rather than the Current setpoint.  The README was updated to reflect this.  In addition, the maxima for the max Current and Temperature inputs were raised to 14 A and 1300 C, respectively.  In addition, the requirement that max temperature be above 0 C was lifted.  The initial values for Kp and Ki were changed.