Max current light functionality changed.
The max current light now goes on when the measured current is within .1A of the user-defined maximum.  This change was necessary because the max current light would go on in the case of max voltage when the setpoint continued to rise, even though the actual current was well below the maximum.  In addition, the tabbing order of the fron panel was changed.  The README was edited to reflect the change in light functionality.