Controller can now read T from LakeShore.
The TPD controller now reads temperature data from the LakeShore temperature controller.  This data comes from a thermocouple connected to the temperature controller.  Data is read every 500 ms, from the same loop as the PID algorithm and the power supply code.  The control algorithm uses this data rather than the T data from the pyrometer because it is more stable and accurate.  However, the data from the pyrometer is still collected and stored as a second resource.  In addition, the temperature schedule is now written into the text file.