Added "release port" VI.
release_port.vi was written in case the COM1 serial port is tied up.  If this is the case, the user can quit LabVIEW and open and run release_port.vi to close any preexisting connections to the port.  In addition, an error in the formatting of the text file was fixed, and a function was added to close the connection to the power supply.