Wrote program to tune PID using open loop.
open-loop_PID_tuner.vi is a tool to manually tune the PID algorithm for TPD_Controller.vi.  It can be used to obtain and save the data necessary to complete open-loop tuning as described in the manual for the LabVIEW PID toolkit.  The user can pass an arbitrary current through the sample, and the program stores the elapsed time [s], Current setpoint [A], measured Current [A], and measured temperature [C] in a text file.  The data is saved in columns delimited by tabs in that order.  The power supply voltage is set to 20 V.
In addition, in TPD_Controller.vi, Td is now saved in the text file.