Simulink simulation updated.
The simulink simulation of the control system is very close now.  I used a piecewise T to R function calculated from data from an open-loop current ramp on the sapphire sample holder.  The simulation can accurately predict the temperature at which a P controller will settle, and begins oscillating at the right time.  The only problem is that the oscillations are much faster than in the real system because the time constant falling temperature in the real system is much higher than for rising temperature.  Kc values from .002 to .015 were tested, and all are somewhat close, except for the frequency of oscillation.