More data added and changes to TPD_Controller.vi
Data was added for the first TPD runs with RGA data being collected.  Also, TPD_Controller.vi was changed.  The code that gets data from teh DAQ board was moved into the same loop as the code that polls the temperature controller and writes to the power supply.  This fixes the issue of the pyrometer temperature only being collected in the first run, and not in subsequent ones.  The pyrometer data is averaged over 5 points.  This means that the pyrometer data is now less responsive and less accurate.